generalize GenericRecipePage.updateButtonsForPage
