+++ common/src/main/java/marsh/town/brb/brewingstand/BrewingRecipeBookResults.java
@@ -2,7 +2,6 @@ package marsh.town.brb.brewingstand;
 
 import marsh.town.brb.BetterRecipeBook;
 import marsh.town.brb.generic.GenericRecipePage;
-import marsh.town.brb.recipe.BRBRecipeBookCategory;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.core.RegistryAccess;
@@ -10,8 +9,6 @@ import net.minecraft.world.inventory.BrewingStandMenu;
 
 @Environment(EnvType.CLIENT)
 public class BrewingRecipeBookResults extends GenericRecipePage<BrewingStandMenu, BrewingRecipeCollection, BrewableResult, BrewableRecipeButton> {
-    BRBRecipeBookCategory category;
-
     public BrewingRecipeBookResults(RegistryAccess registryAccess) {
         super(registryAccess);
 
@@ -27,23 +24,6 @@ public class BrewingRecipeBookResults extends GenericRecipePage<BrewingStandMenu
 
     @Override
     protected void initOverlay(BrewingRecipeCollection recipeCollection, int x, int y, RegistryAccess registryAccess) {
-    }
-
-    public void updateButtonsForPage() {
-        int i = 20 * this.currentPage;
-
-        for (int j = 0; j < this.buttons.size(); ++j) {
-            BrewableRecipeButton brewableRecipeButton = this.buttons.get(j);
-            if (i + j < this.recipeCollections.size()) {
-                BrewingRecipeCollection output = this.recipeCollections.get(i + j);
-                brewableRecipeButton.showCollection(output, menu, category);
-                brewableRecipeButton.visible = true;
-            } else {
-                brewableRecipeButton.visible = false;
-            }
-        }
-
-        this.updateArrowButtons();
     }
 
     @Override
